Take in redirectUrlTemplate to allow the frontend to decide path format
This is useful to not create a dependency between this backend and the
frontend. The frontend can now change the cancellation path to anything
it wants any time, and the mail which is sent out will reflect this.

We do a simple string search and replace and hope that is good enough.

We also return the redirectUrl in case the frontend wants to make use of
it immediately, but  this might be overkill.